requirements.txt
customtkinter
opencv-python
face-recognition
numpy
Pillow
pygame
